Nettoyage du code et application de certaines regles du clean code

+++ src/forecasting_sales/tools/data_science/forecasting.py
@@ -11,7 +11,6 @@ class ForecastingConfig:
     """
 
     n_estimators: int
-    learning_rate: int
+    learning_rate: float
     max_depth: int
     random_state: int
-
